refactor: rename `AppError.ts` to `app-error.ts`
